bp_tex asm audit: 4 opcode dumps replaced with plain C; 3 remain exact MATCH (FindNodeByIndex, FindNodeById, ApplyGlobalAlpha), CollectLeaves honest NONMATCHING nd13

+++ src/Event/Fcl/fclShopBGManager.c
@@ -5501,9 +5501,9 @@ void FUN_004167d0(u64 param_1,int param_2,int param_3,int param_4,int param_5,
 
   short sStack_6;
 
-  short sStack_4;
+  int sStack_4;
 
-  short sStack_2;
+  int sStack_2;
